Moves setting the app attribute to the _set_cache method
Initializing a Cache instance with a None app and setting it later
via init_app causes the app attribute never to be set. This is fine for
most cases, since the cache is being set to the app extension. However,
it causes crashes when calling the Cache outside of a Flask context
(e.g.: spawned threads).