Adding source check functionality to memoize.
By default source check is disabled (source_check=False) which
preserves existing behaviour to use the cached value if the source
code changes for the function.

By enabling source check, the function's source code will be included
when forming the hash which will be used as a key to store the value.